# Improve performance of addIndexedMappings