This does two things:

1. Does not normalize path separators except on windows since this is expensive.
2. Implements `addIndexedMappings` natively for napi so that it is a single native call instead of one for each mapping. This is done by encoding all mappings into a buffer to pass them to C++.

Before:

```
consume JS Mappings 6404.88 opts/sec (mean: 0.156ms, stddev: 0.122ms, 250 samples)
```

After:

```
consume JS Mappings 20021.29 opts/sec (mean: 0.1ms, stddev: 0.193ms, 250 samples)
```

This improves overall build perf for three.js by ~3%.