@DeMoorJasper 2020-12-13
Experimented with entirely removing strings from this and added another 2x faster processing of addIndexedMappings, strings are so slow.

Will also try and port as much of this as possible to WebAssembly.